Drain KernelStep starts and completions independently and abandon lost completions.
A KernelStep completion still missing after KERNEL_STEP_COMPLETION_LOST_LAG
later starts on the same channel is closed so it no longer freezes later
stops or leaves CoMMA holding a start-only handle. Also publishes KernelStep
support to the profiler plugin, drains children before KernelCh stop, links
libibverbs through LDFLAGS, and falls back when ibv_query_port_speed is absent.

+++ src/device/sendrecv.h
@@ -107,7 +107,7 @@ struct RunWorkBatch<ncclFuncSendRecv, T, RedOp, NCCL_ALGO_RING, NCCL_PROTO_SIMPL
         if ((profActive >> wi) & 1) {
           int localIx = __popc(profActive & ((1u << wi) - 1u));
           works[wi].profilerWorkTag =
-            (uint16_t)(ncclShmem.channel.workCounter + localIx + 1);
+            (uint32_t)(ncclShmem.channel.workCounter + localIx + 1);
         } else {
           works[wi].profilerWorkTag = 0;
         }
